staging: mmal-vchiq: Fix client_component for 64 bit kernel
BugLink: https://bugs.launchpad.net/bugs/1982090

The MMAL client_component field is used with the event
mechanism to allow the client to identify the component for
which the event is generated.
The field is only 32bits in size, therefore we can't use a
pointer to the component in a 64 bit kernel.

Component handles are already held in an array per VCHI
instance, so use the array index as the client_component handle
to avoid having to create a new IDR for this purpose.

Signed-off-by: Dave Stevenson <dave.stevenson@raspberrypi.org>

(cherry picked from commit f2d332d1d3ded18f1f57c9b0e20eb38ce0b8a9bc rpi-5.19.y)
Signed-off-by: Juerg Haefliger <juerg.haefliger@canonical.com>