bcm2835-codec: Return empty buffers to the VPU instead of queueing to vbuf2
BugLink: https://bugs.launchpad.net/bugs/1982090

The encoder can skip frames totally should rate control overshoot
the target bitrate too far. In this situation it generates an
output buffer of length 0.
V4L2 treats a buffer of length 0 as an end of stream flag, which is
not appropriate in this case, therefore we can not return that buffer
to the client.

The driver was returning the buffer to videobuf2 in the QUEUED state,
however that buffer was then not dequeued again, so the number of
buffers was reduced each time this happened. In the pathological
case of using GStreamer's videotestsrc in mode 1 for noise, this happens
sufficiently frequently to totally stall the pipeline.

If the port is still enabled then return the buffer straight back to
the VPU rather than to videobuf2.

Signed-off-by: Dave Stevenson <dave.stevenson@raspberrypi.com>

(cherry picked from commit 1eebb16b88eef028c126514fbeb1343f98499720 rpi-5.19.y)
Signed-off-by: Juerg Haefliger <juerg.haefliger@canonical.com>