drm/vc4: Force trigger of dlist update on margins change
BugLink: https://bugs.launchpad.net/bugs/1982090

When the margins are changed, the dlist needs to be regenerated
with the changed updated dest regions for each of the planes.

Setting the zpos_changed flag is sufficient to trigger that
without doing a full modeset, therefore set it should the
margins be changed.

Signed-off-by: Dave Stevenson <dave.stevenson@raspberrypi.com>

(cherry picked from commit 872504f4767af6bcc4953e8336013c8764341df8 rpi-5.19.y)
Signed-off-by: Juerg Haefliger <juerg.haefliger@canonical.com>