media: bcm2835-unicam: Correctly handle FS + FE ISR condtion
BugLink: https://bugs.launchpad.net/bugs/1989958

If we get a simultaneous FS + FE interrupt for the same frame, it cannot be
marked as completed and returned to userland as the framebuffer will be refilled
by Unicam on the next sensor frame. Additionally, the timestamp will be set to 0
as the FS interrupt handling code will not have run yet.

To avoid these problems, the frame is considered dropped in the FE handler,
and will be returned to userland on the subsequent sensor frame.

Signed-off-by: Naushir Patuck <naush@raspberrypi.com>

(cherry picked from commit c3a3eb5a35904fd4da0a2f930b0285520577fc8e rpi-5.19.y)
Signed-off-by: Juerg Haefliger <juerg.haefliger@canonical.com>